src/main/java/kr/megaptera/assignment/application/CreatePostService.java
package kr.megaptera.assignment.application;

import kr.megaptera.assignment.dtos.PostDto;
import kr.megaptera.assignment.models.Post;
import kr.megaptera.assignment.repositories.PostRepository;
import org.springframework.stereotype.Service;

@Service
public class CreatePostService {
    private final PostRepository postRepository;

    public CreatePostService(PostRepository postRepository) {
        this.postRepository = postRepository;
    }

    public PostDto createPost(PostDto postDto) {
        Post newPost = new Post(postDto.getTitle(),
                postDto.getAuthor(),
                postDto.getContent());
        postRepository.save(newPost);
        return new PostDto(newPost);
    }
}
